delete EnrollmentDisplay wrapper; consume dashboards directly
HomeContent now imports HomeEnrollmentsDashboard; ProgramContent
imports ProgramEnrollmentDisplay. The EnrollmentDisplay dispatch
wrapper (programId ? Program : Home) added no value once consumers
already knew which dashboard they wanted.

EnrollmentDisplay.tsx renamed to ProgramEnrollmentDisplay.tsx (the
remaining content). Its test file follows; describe collapses from
EnrollmentDisplay > ProgramId Filtering > tests to ProgramEnrollmentDisplay.

ProgramAsCourseCard's ProgramCertificateButton import path updates;
DashboardDialogs.test.tsx switches its home-flow render to
HomeEnrollmentsDashboard directly.
